ev-open-telemetry/ev_observer/pyproject.toml
[tool.poetry]
name = "tesla-to-greptime"
version = "0.1.0"
description = ""
authors = ["truedrju <76677784+truedrju@users.noreply.github.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.9"
teslapy = "^2.9.0"
python-dotenv = "^1.0.1"
opentelemetry-api = "^1.27.0"
opentelemetry-exporter-otlp-proto-common = "^1.27.0"
opentelemetry-exporter-otlp-proto-http = "^1.27.0"
opentelemetry-instrumentation = "^0.48b0"
opentelemetry-instrumentation-system-metrics = "^0.48b0"
opentelemetry-proto = "^1.27.0"
opentelemetry-sdk = "^1.27.0"
pydantic = "^2.9.2"


[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "E266", "E501", "W503"]
max-complexity = 18
select = ["B", "C", "E", "F", "W", "T4", "B9"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
